add: move component classes to new file
